Extract control bar related code into seperate module
